+++ server/models/index.js
@@ -1,43 +1,30 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
-const basename = path.basename(__filename);
-const db = {};
-
-const sequelize = new Sequelize('db_bid_local', 'db_user', 'BID_local@2020', {
-  host: 'localhost',
-  dialect: 'mysql',
-  logging: false,
-  pool: {
-    max: 5,
-    min: 0,
-    acquire: 30000,
-    idle: 10000
+let users = {
+  1: {
+    id: '1',
+    username: 'Robin Wieruch',
+    itemIds: [1],
   },
-  operatorsAliases: false // http://docs.sequelizejs.com/manual/tutorial/querying.html#operators
-});
-
-fs
-  .readdirSync(__dirname)
-  .filter(file => {
-    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
-  })
-  .forEach(file => {
-    // var model = sequelize['import'](path.join(__dirname, file));
-    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
-    db[model.name] = model;
-  });
-
-Object.keys(db).forEach(modelName => {
-  console.log(modelName);
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
-});
-
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
-
-module.exports = db;
+  2: {
+    id: '2',
+    username: 'Dave Davids',
+    itemIds: [2],
+  },
+};
+ 
+let items = {
+  1: {
+    id: '1',
+    name: 'Hello World',
+    userId: '1',
+  },
+  2: {
+    id: '2',
+    name: 'By World',
+    userId: '2',
+  },
+};
+ 
+module.exports =  {
+  users,
+  items
+};
